+++ src/guy.cpp
@@ -18,8 +18,8 @@
 #include <math.h>
 #define RAISE 1.85  // please also change in picker.cpp
 
-extern long costlist[NUM_FAMILIES];  // These come from picker.cpp
-extern long statcosts[NUM_FAMILIES][6];
+extern int costlist[NUM_FAMILIES];  // These come from picker.cpp
+extern int statcosts[NUM_FAMILIES][6];
 // Zardus: PORT, exception doesn't compile (dos thing?): int matherr(struct exception *);
 
 int MAX(int a,int b)
@@ -245,10 +245,10 @@ guy::~guy()
 	next = NULL;
 }
 
-long guy::query_heart_value() // how much are we worth?
+int guy::query_heart_value() // how much are we worth?
 {
 	guy *normal = new guy(family); // for base comparisons
-	long cost=0, temp;
+	int cost=0, temp;
 
 	if (!normal)
 		return 0l;
@@ -256,35 +256,35 @@ long guy::query_heart_value() // how much are we worth?
 	// Get strength cost ..
 	temp = strength - normal->strength; // difference..
 	temp = MAX(temp,0);
-	cost += (long) (pow( temp, RAISE)
-	                * (long)statcosts[(int)family][0]);
+	cost += (int) (pow( temp, RAISE)
+	                * (int)statcosts[(int)family][0]);
 
 	// Get dexterity cost ..
 	temp = dexterity - normal->dexterity; // difference..
 	temp = MAX(temp,0);
-	cost += (long) (pow( temp, RAISE)
-	                * (long)statcosts[(int)family][1]);
+	cost += (int) (pow( temp, RAISE)
+	                * (int)statcosts[(int)family][1]);
 
 	// Get constitution cost ..
 	temp = constitution - normal->constitution; // difference..
 	temp = MAX(temp,0);
-	cost += (long) (pow( temp, RAISE)
-	                * (long)statcosts[(int)family][2]);
+	cost += (int) (pow( temp, RAISE)
+	                * (int)statcosts[(int)family][2]);
 
 	// Get intelligence cost ..
 	temp = intelligence - normal->intelligence; // difference..
 	temp = MAX(temp,0);
-	cost += (long) (pow( temp, RAISE)
-	                * (long)statcosts[(int)family][3]);
+	cost += (int) (pow( temp, RAISE)
+	                * (int)statcosts[(int)family][3]);
 
 	// Get armor cost ..
 	temp = armor - normal->armor; // difference..
 	temp = MAX(temp,0);
-	cost += (long) (pow( temp, RAISE)
-	                * (long)statcosts[(int)family][4]);
+	cost += (int) (pow( temp, RAISE)
+	                * (int)statcosts[(int)family][4]);
 
 	// Add in the base cost value for the guy ..
-	cost += (long) costlist[(int)family];
+	cost += (int) costlist[(int)family];
 
 	return cost;
